+++ frontend/src/components/Trials/Selection/SelectionOneCard.tsx
